requirements.txt
alembic==1.8.1
anyio==3.6.2
attrs==22.1.0
bcrypt==4.0.1
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.86.0
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
mypy-extensions==0.4.3
orjson==3.8.1
packaging==22.0
passlib==1.7.4
pathspec==0.10.1
platformdirs==2.5.3
pluggy==1.0.0
psycopg2==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pytest==7.2.0
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a29
sqlmodel==0.0.8
starlette==0.20.4
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
